feat(bulk-data): adapt bulk data generation to set admin as owner of all assets + move import bulk data function outside of the sevice file + beging template feature
