+++ plugins/leemons-plugin-bulk-data/backend/core/importHandlers/index.js
@@ -1,0 +1,5 @@
+const { getLoadStatus } = require('./importBulkData');
+const { loadFromFile } = require('./loadFromFile');
+const { loadFromTemplateURL } = require('./loadTemplateFromUrl');
+
+module.exports = { loadFromFile, getLoadStatus, loadFromTemplateURL };
